+++ libraries/OUTPIN/CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.1] - 2026-01-12
+- add **void multiPulseHigh(uint8_t nopsHIGH, uint8_t nopsLOW, uint8_t times)**
+- add **void multiPulseLow(uint8_t nopsHIGH, uint8_t nopsLOW, uint8_t times)**
+- update GitHub actions
+- update examples
+- minor edits
+
 ## [0.1.0] - 2025-02-06
 - initial release